fix a bug where some scripts fail
