+++ scripts/train.py
@@ -10,8 +10,8 @@ from omegaconf import DictConfig, ListConfig
 from pytorch_lightning.callbacks import Callback
 from pytorch_lightning.loggers import Logger
 
-from kwja.cli.utils import filter_logs
 from kwja.datamodule.datamodule import DataModule
+from kwja.utils.logging import filter_logs
 
 filter_logs(environment="development")
